# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Forum API
  description: REST API for forum posts and users (JWT bearer auth).
  version: 1.0.0
servers:
  - url: http://localhost:9000
    description: Local
...
paths:
...
  /posts/{postId}/replies:
    get:
      tags: [Replies]
      operationId: listRepliesForPost
      summary: List replies for a post
      security: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReplyResponse"
        "404":
          description: Post not found
  /replies:
    post:
      tags: [Replies]
      operationId: createReply
      summary: Create a reply on a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReplyRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyResponse"
        "404":
          description: Post not found
  /replies/{id}:
    get:
      tags: [Replies]
      operationId: getReplyById
      summary: Get reply by id
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyResponse"
        "404":
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateReplyRequest:
      type: object
      required: [content, userId]
      properties:
        content:
          type: string
          maxLength: 10000
        userId:
          type: integer
          format: int64
        postId:
          type: integer
          format: int64
    ReplyResponse:
      type: object
      required: [id, postId, content, createdAt]
      properties:
        id:
          type: integer
          format: int64
        postId:
          type: integer
          format: int64
        content:
          type: string
        createdAt:
          type: string
          format: date-time
security:
  - bearerAuth: []